server/requirements.txt
fastapi==0.115.9
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv
